CooperativeTasks

Documentation for CooperativeTasks.

CooperativeTasks.ALIVE

CooperativeTasks.DEAD

CooperativeTasks.ESTABLISHED

CooperativeTasks.INITIALIZATION

CooperativeTasks.LOST

CooperativeTasks.SYNC_VARNAME

CooperativeTasks.UNRESPONSIVE

CooperativeTasks.Command

CooperativeTasks.Connection

CooperativeTasks.Discard

CooperativeTasks.LoopExecution

CooperativeTasks.SingleExecution

CooperativeTasks.SpawnOptions

CooperativeTasks.StatusCode

CooperativeTasks.TaskGroup

CooperativeTasks.TaskState

CooperativeTasks.remove_child

CooperativeTasks.shutdown

CooperativeTasks.shutdown_children

CooperativeTasks.spawn

CooperativeTasks.trysend

CooperativeTasks.@check

CooperativeTasks.ALIVE — Constant

Communication is theoretically possible with the task.

CooperativeTasks.DEAD — Constant

The task is no longer running; either it has signalled its death or it is marked as done.

CooperativeTasks.ESTABLISHED — Constant

The connection has been established and is nominal.

CooperativeTasks.INITIALIZATION — Constant

The connection has been initialized.

CooperativeTasks.LOST — Constant

The connection was established then lost, due to a failure from either end.

CooperativeTasks.SYNC_VARNAME — Constant

Variable used for Base.@sync blocks.

CooperativeTasks.UNRESPONSIVE — Constant

The task has timed out once or more on recent requests and has not shown activity since then.

CooperativeTasks.Command — Type

Execute ret = f() on a task, optionally executing continuation(ret) from the task the message has been sent from.

First, the command is registered on a source task with a corresponding UUID. Then, as part of a message, it is sent to the destination task for execution, which will send back the value associated with this UUID if any continuation has been provided. If so, when the source task next collects new messages, it will run continuation with the returned value.

CooperativeTasks.Connection — Type

Connection established with a peer, from a task-local perspective.

CooperativeTasks.Discard — Type

Placeholder for a Future UUID to prevent any value from being stored for this Future.

CooperativeTasks.LoopExecution — Type

Execute a function repeatedly until a shutdown is scheduled with schedule_shutdown.

If period is greater than a millisecond, every iteration may trigger a sleep for the remaining period time after executing the main function. Simple heuristics are used to prevent sleeping when there were recent interactions with other tasks, to avoid suffering from large communication delays.

At every iteration, task messages will be checked and allow (among other things) to compute/return results to other tasks (see Future), and to cancel the task. This mode of execution is preferred over manual loops precisely for the ability to satisfy task duties as required in the context of this library.

CooperativeTasks.SingleExecution — Type

Execute a function once, then return.

CooperativeTasks.SpawnOptions — Type

Options for spawning tasks with spawn.

CooperativeTasks.StatusCode — Type

Status code used for communicating success, failures and possibly other intermediate return modes.

CooperativeTasks.TaskGroup — Type

List of tasks to be shutdown via a finalizer once the lifetime of TaskGroup expires.

CooperativeTasks.TaskState — Type

Task state hinting at whether it should or may not answer to future requests.

CooperativeTasks.remove_child — Method

Request a parent task to remove the current task as child.

CooperativeTasks.shutdown — Method

Shut down a task by cancelling it if it has not completed.

See cancel.

CooperativeTasks.shutdown_children — Method

Shutdown all children of the current task. Returns a Condition which can be waited on.

CooperativeTasks.spawn — Method

Spawn the function f on a new task.

See also: SpawnOptions

CooperativeTasks.trysend — Method

Send a message m to task.

CooperativeTasks.@check — Macro

@check f(args...)

Assign the expression to a variable named _return_code. Then, if the value is not a success code, return a ConcurrencyError holding the return code.