ConcurrencyGraph

Documentation for ConcurrencyGraph.

ConcurrencyGraph.ALIVE

ConcurrencyGraph.DEAD

ConcurrencyGraph.ESTABLISHED

ConcurrencyGraph.INITIALIZATION

ConcurrencyGraph.LOST

ConcurrencyGraph.UNRESPONSIVE

ConcurrencyGraph.Command

ConcurrencyGraph.Connection

ConcurrencyGraph.Discard

ConcurrencyGraph.StatusCode

ConcurrencyGraph.TaskGroup

ConcurrencyGraph.TaskState

ConcurrencyGraph.remove_child

ConcurrencyGraph.send

ConcurrencyGraph.shutdown

ConcurrencyGraph.shutdown_children

ConcurrencyGraph.@check

ConcurrencyGraph.ALIVE — Constant

Communication is theoretically possible with the task.

ConcurrencyGraph.DEAD — Constant

The task is no longer running; either it has signalled its death or it is marked as done.

ConcurrencyGraph.ESTABLISHED — Constant

The connection has been established and is nominal.

ConcurrencyGraph.INITIALIZATION — Constant

The connection has been initialized.

ConcurrencyGraph.LOST — Constant

The connection was established then lost, due to a failure from either end.

ConcurrencyGraph.UNRESPONSIVE — Constant

The task has timed out once or more on recent requests and has not shown activity since then.

ConcurrencyGraph.Command — Type

Execute ret = f() on a task, optionally executing continuation(ret) from the task the message has been sent from.

First, the command is registered on a source task with a corresponding UUID. Then, as part of a message, it is sent to the destination task for execution, which will send back the value associated with this UUID if any continuation has been provided. If so, when the source task next collects new messages, it will run continuation with the returned value.

ConcurrencyGraph.Connection — Type

Connection established with a peer, from a task-local perspective.

ConcurrencyGraph.Discard — Type

Placeholder for a Future UUID to prevent any value from being stored for this Future.

ConcurrencyGraph.StatusCode — Type

Status code used for communicating success, failures and possibly other intermediate return modes.

ConcurrencyGraph.TaskGroup — Type

List of tasks to be shutdown via a finalizer once the lifetime of TaskGroup expires.

ConcurrencyGraph.TaskState — Type

Task state hinting at whether it should or may not answer to future requests.

ConcurrencyGraph.remove_child — Method

Request a parent task to remove the current task as child.

ConcurrencyGraph.send — Method

Send a message m to task.

ConcurrencyGraph.shutdown — Method

Shut down a task by cancelling it if it has not completed.

See cancel.

ConcurrencyGraph.shutdown_children — Method

Shutdown all children of the current task. Returns a Condition which can be waited on.

ConcurrencyGraph.@check — Macro

@check f(args...)

Assign the expression to a variable named _return_code. Then, if the value is not a success code, return a ConcurrencyError holding the return code.