ConcurrencyGraph

Documentation for ConcurrencyGraph.

ConcurrencyGraph.Command

ConcurrencyGraph.TaskGroup

ConcurrencyGraph.send

ConcurrencyGraph.shutdown

ConcurrencyGraph.Command — Type

Execute ret = f() on a task, optionally executing continuation(ret) from the task the message has been sent from.

First, the command is registered on a source task with a corresponding UUID. Then, as part of a message, it is sent to the destination task for execution, which will send back the value associated with this UUID if any continuation has been provided. If so, when the source task next collects new messages, it will run continuation with the returned value.

ConcurrencyGraph.TaskGroup — Type

List of tasks to be shutdown via a finalizer once the lifetime of TaskGroup expires.

ConcurrencyGraph.send — Method

Send a message m to task.

ConcurrencyGraph.shutdown — Method

Shut down a task by cancelling it if it has not completed.

See cancel.